preinstall: Reduce some code duplication
This moves the code to read the current boot load option from the log
into a helper function, as the same code was used both in the PCR4 and
PCR7 checks.
